fix: satisfy backend-size and pkg-file-count for work-family PR
Refresh contract staging manifest sourceCommit, shrink climanifestgen test via testdata fixture, and relocate session constructors into root_workflow.go so root_work.go stays under the 1000-line backend-size limit without exceeding the 15-file pkg/transports/cli cap.
